feat: using a new analyzer
