+++ src/server/index.ts
@@ -17,6 +17,9 @@ async function openBrowser(address: string) {
 // Don't worry we'll still respecet user options
 // Or we also want user to use the plugin on rollup.
 function hijackViteMinifyPlugin(conf: ConfigResolveParameter) {
+  // https://vitejs.dev/config/build-options.html#build-sourcemap
+  // ensure `sourcemap` option
+  if (!conf.build.sourcemap) conf.build.sourcemap = 'hidden'
   // https://github.com/vitejs/vite/blob/716286ef21f4d59786f21341a52a81ee5db58aba/packages/vite/src/node/plugins/esbuild.ts#L289-L307
   const internalPluginNames = ['vite:esbuild-transpile', 'vite:terser']
   // Minify logic
@@ -24,9 +27,6 @@ function hijackViteMinifyPlugin(conf: ConfigResolveParameter) {
   const { minify } = conf.build
   const name = [true, 'esbuild'].includes(minify) ? internalPluginNames[0] : minify === 'terser' ? internalPluginNames[1] : ''
   if (!name) return 
-  // https://vitejs.dev/config/build-options.html#build-sourcemap
-  // ensure `sourcemap` option
-  if (!conf.build.sourcemap) conf.build.sourcemap = 'hidden'
   const plugin = conf.plugins.find(p => p.name === name)
   const hook = plugin?.renderChunk
   if (!hook) return
@@ -56,6 +56,7 @@ function analyzer(opts: AnalyzerPluginOptions = { analyzerMode: 'server' }): Plu
       hijackViteMinifyPlugin(config)
     },
     async generateBundle(_, outputBundle) {
+      analyzerModule.installPluginContext(this)
       // After consider. I trust process chunk is enougth. (If you don't think it's right. PR welcome.)
       for (const bundleName in outputBundle) {
         const bundle = outputBundle[bundleName]
@@ -70,19 +71,19 @@ function analyzer(opts: AnalyzerPluginOptions = { analyzerMode: 'server' }): Plu
       switch (opts.analyzerMode) {
         case 'json': {
           const p = path.join(defaultWd, opts.fileName ? `${opts.fileName}.json` : 'stats.json')
-          const foamModule = await analyzerModule.processFoamModule()
+          const foamModule = analyzerModule.processFoamModule()
           fsp.writeFile(p, JSON.stringify(foamModule, null, 2), 'utf8')
           break
         }
         case 'static': {
           const p = path.join(defaultWd, opts.fileName ? `${opts.fileName}.html` : 'stats.html')
-          const foamModule = await analyzerModule.processFoamModule()
+          const foamModule = analyzerModule.processFoamModule()
           const html = await renderView(foamModule, { title: reportTitle, mode: 'stat' })
           fsp.writeFile(p, html, 'utf8')
           break
         }
         case 'server': {
-          const foamModule = await analyzerModule.processFoamModule()
+          const foamModule = analyzerModule.processFoamModule()
           const { setup, port } = createServer((opts.analyzerPort === 'atuo' ? 0 : opts.analyzerPort) ?? 8888)
           setup(foamModule, { title: reportTitle, mode: 'stat' })
           if ((opts.openAnalyzer ?? true) && !isCI) {
